Move EIP code into eip.py, and add some tests
